// src/app/model/APIModel/OnlineUser.ts
/**
 * Created by zhaoxinlei on 2018/3/23.
 */
export class OnlineUser {
  Id: string;
  Identify: string;
  IdentifyType: string;
  IP: string;
  LoginTime: string;
  MacAddr: string;
  Message: string;
  Online: boolean;
  OrgList: any[];
  Password: string;
  PlatCustomerId: string;
  PrivId: string;
  ProjId: string;
  ProjList: any[];
  TryTimes: number;
  RemainTimes: number;
  UserId: string;
  ValidCode: string;
  ValidCodeId: string;
  Token: string;
}